Added Java/JVM and JavaScript development tool installation scripts

+++ install/jetbrains_toolbox.sh
@@ -1,22 +1,37 @@
 #!/bin/bash
+set -euo pipefail
 
 # Installer for JetBrains Toolbox
 # This will allow the installation, upgrade and management of other JetBrains tools
 # Running the tool itself will automatically add an icon to the launcher
 
 VERSION=1.1.2143
-TMPDIR=~/tools/tmp
-OUTDIR=~/tools/jetbrains
+TMPDIR=$HOME/tools/tmp
+OUTDIR=$HOME/tools/jetbrains
+CHECKSUM="74ca89a1b97367e909075e53c67ee093ca316ecb8124bb39e7318750634552a3"
 
-mkdir -p $OUTDIR
-mkdir -p $TMPDIR
+if [ ! -x "$OUTDIR/jetbrains-toolbox-$VERSION" ]; then
+    mkdir -p $TMPDIR
 
-curl -L https://download.jetbrains.com/toolbox/jetbrains-toolbox-${VERSION}.tar.gz -o $TMPDIR/jetbrains-toolbox.tar.gz
-cd $OUTDIR
-tar xvzf $TMPDIR/jetbrains-toolbox.tar.gz
-rm -f $OUTDIR/latest
-ln -s $OUTDIR/*$VERSION* $OUTDIR/latest
+    curl -L https://download.jetbrains.com/toolbox/jetbrains-toolbox-${VERSION}.tar.gz -o $TMPDIR/jetbrains-toolbox.tar.gz
 
-rm -r $TMPDIR
+    if [ "$(sha256sum $TMPDIR/jetbrains-toolbox.tar.gz)" == "$CHECKSUM  $TMPDIR/jetbrains-toolbox.tar.gz" ]; then
+        mkdir -p $OUTDIR
+        cd $OUTDIR
+        tar xvzf $TMPDIR/jetbrains-toolbox.tar.gz
+        rm -f $OUTDIR/latest
+        ln -s $OUTDIR/jetbrains-toolbox-$VERSION $OUTDIR/latest
+    else
+        echo "Invalid checksum for tarball.  Aborting."
+        exit 1
+    fi
 
-$OUTDIR/latest/jetbrains-toolbox &
+    rm -r $TMPDIR
+
+    $OUTDIR/latest/jetbrains-toolbox &
+else
+    echo "JetBrains Toolbox $VERSION is already installed"
+    echo "To remove it and re-install run rm -rf $OUTDIR/jetbrains-toolbox-$VERSION"
+fi
+
+exit 0
